refactor: move delegation revoke UI into delegation domain
