+++ src/features/delegation/screens/ViewWalletDelegations.tsx
@@ -11,26 +11,26 @@ import ContextMenuButton from '@/components/native-context-menu/contextMenu';
 import { Box, Separator, Stack, Text, useForegroundColor } from '@/design-system';
 import { fonts } from '@/design-system/typography/typography';
 import { IS_DEV } from '@/env';
-import {
-  hasActiveDelegation,
-  isRainbowDelegated as hasRainbowDelegation,
-  isThirdPartyDelegated as hasThirdPartyDelegation,
-} from '@/features/delegation/utils/status';
 import { ChainImage } from '@/features/network/components/ChainImage';
 import { backendNetworksActions } from '@/features/network/stores/backendNetworksStore';
 import { type ChainId } from '@/features/network/types/backendNetworks';
 import * as i18n from '@/languages';
 import { navigate } from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
-import { RevokeReason } from '@/screens/delegation/RevokeDelegationPanel';
+import Menu from '@/screens/SettingsSheet/components/Menu';
+import MenuContainer from '@/screens/SettingsSheet/components/MenuContainer';
+import MenuItem from '@/screens/SettingsSheet/components/MenuItem';
 import { useTheme } from '@/theme/ThemeContext';
 import { formatAddressForDisplay } from '@/utils/abbreviations';
 import * as ethereumUtils from '@/utils/ethereumUtils';
 import { delegation, useDelegationDisabled, useDelegations, type DelegationWithChainId } from '@rainbow-me/delegation';
 
-import Menu from '../Menu';
-import MenuContainer from '../MenuContainer';
-import MenuItem from '../MenuItem';
+import {
+  hasActiveDelegation,
+  isRainbowDelegated as hasRainbowDelegation,
+  isThirdPartyDelegated as hasThirdPartyDelegation,
+} from '../utils/status';
+import { RevokeReason } from './RevokeDelegationPanel';
 
 type ViewWalletDelegationsParams = {
   ViewWalletDelegations: { walletId: string; address: Address; title: string };
