+++ src/navigation/Routes.android.tsx
@@ -16,6 +16,7 @@ import { useExperimentalFlag } from '@/features/config/hooks/experimentalHooks';
 import { SignTransactionSheet } from '@/features/dapp-request/screens/SignTransactionSheet';
 import { LogSheet } from '@/features/debug/screens/LogSheet';
 import { WalletDiagnosticsSheet } from '@/features/debug/screens/WalletDiagnosticsSheet';
+import { RevokeDelegationPanel } from '@/features/delegation/screens/RevokeDelegationPanel';
 import RegisterENSNavigator from '@/features/ens/navigation/RegisterENSNavigator';
 import ENSAdditionalRecordsSheet from '@/features/ens/screens/ENSAdditionalRecordsSheet';
 import ENSConfirmRegisterSheet from '@/features/ens/screens/ENSConfirmRegisterSheet';
@@ -67,7 +68,6 @@ import { ActivitySheetScreen } from '@/screens/ActivitySheetScreen';
 import { AirdropsSheet } from '@/screens/Airdrops/AirdropsSheet';
 import { ClaimAirdropSheet } from '@/screens/Airdrops/ClaimAirdropSheet';
 import { ClaimClaimablePanel } from '@/screens/claimables/ClaimPanel';
-import { RevokeDelegationPanel } from '@/screens/delegation/RevokeDelegationPanel';
 import { ExpandedAssetSheet as ExpandedAssetSheetV2 } from '@/screens/expandedAssetSheet/ExpandedAssetSheet';
 import LearnWebViewScreen from '@/screens/LearnWebViewScreen';
 import MintSheet from '@/screens/mints/MintSheet';
